added loadConstraints (check comment)
the container is hidden by default, and only after both:
image is loaded (if it's enabled)
layout from localStorage is applied
only then show the container

this is because sometimes you could either see the default layout
flash for one frame, or you could see the half-loaded image.

this fixes both of those.
it does add a 16ms fade transition,
but i think it looks better than bright flash.
you're free to remove it in style.css
under .container